# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Travel Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Visit {
  id: number;
  country: string;
  /** ISO 3166-1 alpha-2 code */
  countryCode: string;
  city: string;
  visitDate: string;
  /** @nullable */
  notes?: string | null;
  lat: number;
  lng: number;
  createdAt: string;
}

export interface VisitInput {
  /** @minLength 1 */
  country: string;
  /**
     * @minLength 2
     * @maxLength 2
     */
  countryCode: string;
  /** @minLength 1 */
  city: string;
  visitDate: string;
  notes?: string;
  lat: number;
  lng: number;
}

export interface VisitUpdate {
  /** @minLength 1 */
  country?: string;
  /**
     * @minLength 2
     * @maxLength 2
     */
  countryCode?: string;
  /** @minLength 1 */
  city?: string;
  visitDate?: string;
  /** @nullable */
  notes?: string | null;
  lat?: number;
  lng?: number;
}

export interface Overview {
  totalVisits: number;
  totalCountries: number;
  totalCities: number;
  /** @nullable */
  topCountry: string | null;
  /** @nullable */
  firstVisit: string | null;
  /** @nullable */
  latestVisit: string | null;
}

export interface YearSummary {
  year: number;
  countries: number;
  cities: number;
  visits: number;
  countryList?: string[];
  cityList?: string[];
}

export interface ErrorResponse {
  error: string;
}

export type ListVisitsParams = {
/**
 * @nullable
 */
year?: number | null;
/**
 * @nullable
 */
country?: string | null;
};