# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Travel Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: visits
    description: Travel visit operations
  - name: stats
    description: Statistics and summary operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /visits:
    get:
      operationId: listVisits
      tags: [visits]
      summary: List all visits
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: country
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of visits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Visit"
    post:
      operationId: createVisit
      tags: [visits]
      summary: Record a new visit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /visits/{id}:
    get:
      operationId: getVisit
      tags: [visits]
      summary: Get a single visit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Visit details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateVisit
      tags: [visits]
      summary: Update a visit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitUpdate"
      responses:
        "200":
          description: Updated visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteVisit
      tags: [visits]
      summary: Delete a visit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats/overview:
    get:
      operationId: getOverview
      tags: [stats]
      summary: Get overall travel stats
      responses:
        "200":
          description: Overview statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Overview"

  /stats/years:
    get:
      operationId: getYearStats
      tags: [stats]
      summary: Get per-year breakdown of visits
      responses:
        "200":
          description: Year summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/YearSummary"

  /stats/timeline:
    get:
      operationId: getTimeline
      tags: [stats]
      summary: Get visits ordered chronologically for map animation
      responses:
        "200":
          description: Chronological visit list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Visit"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Visit:
      type: object
      required: [id, country, countryCode, city, visitDate, lat, lng, createdAt]
      properties:
        id:
          type: integer
        country:
          type: string
        countryCode:
          type: string
          description: ISO 3166-1 alpha-2 code
        city:
          type: string
        visitDate:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        lat:
          type: number
        lng:
          type: number
        createdAt:
          type: string
          format: date-time

    VisitInput:
      type: object
      required: [country, countryCode, city, visitDate, lat, lng]
      properties:
        country:
          type: string
          minLength: 1
        countryCode:
          type: string
          minLength: 2
          maxLength: 2
        city:
          type: string
          minLength: 1
        visitDate:
          type: string
          format: date
        notes:
          type: string
        lat:
          type: number
        lng:
          type: number

    VisitUpdate:
      type: object
      properties:
        country:
          type: string
          minLength: 1
        countryCode:
          type: string
          minLength: 2
          maxLength: 2
        city:
          type: string
          minLength: 1
        visitDate:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        lat:
          type: number
        lng:
          type: number

    Overview:
      type: object
      required: [totalVisits, totalCountries, totalCities, topCountry, firstVisit, latestVisit]
      properties:
        totalVisits:
          type: integer
        totalCountries:
          type: integer
        totalCities:
          type: integer
        topCountry:
          type: ["string", "null"]
        firstVisit:
          type: ["string", "null"]
          format: date
        latestVisit:
          type: ["string", "null"]
          format: date

    YearSummary:
      type: object
      required: [year, countries, cities, visits]
      properties:
        year:
          type: integer
        countries:
          type: integer
        cities:
          type: integer
        visits:
          type: integer
        countryList:
          type: array
          items:
            type: string
        cityList:
          type: array
          items:
            type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string