# lib/api-zod/src/generated/types/overview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Travel Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Overview {
  totalVisits: number;
  totalCountries: number;
  totalCities: number;
  /** @nullable */
  topCountry: string | null;
  /** @nullable */
  firstVisit: Date | null;
  /** @nullable */
  latestVisit: Date | null;
}