# lib/api-zod/src/generated/types/visit.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Travel Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Visit {
  id: number;
  country: string;
  /** ISO 3166-1 alpha-2 code */
  countryCode: string;
  city: string;
  visitDate: Date;
  /** @nullable */
  notes?: string | null;
  lat: number;
  lng: number;
  createdAt: Date;
}