# lib/api-zod/src/generated/types/visitInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Travel Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface VisitInput {
  /** @minLength 1 */
  country: string;
  /**
     * @minLength 2
     * @maxLength 2
     */
  countryCode: string;
  /** @minLength 1 */
  city: string;
  visitDate: Date;
  notes?: string;
  lat: number;
  lng: number;
}