# lib/api-zod/src/generated/types/visitUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Travel Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface VisitUpdate {
  /** @minLength 1 */
  country?: string;
  /**
     * @minLength 2
     * @maxLength 2
     */
  countryCode?: string;
  /** @minLength 1 */
  city?: string;
  visitDate?: Date;
  /** @nullable */
  notes?: string | null;
  lat?: number;
  lng?: number;
}