# lib/api-zod/src/generated/types/yearSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Travel Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface YearSummary {
  year: number;
  countries: number;
  cities: number;
  visits: number;
  countryList?: string[];
  cityList?: string[];
}